Refactor imports to move MessageStatus to libs.types for improved organization
