+++ integrations/slack/parser.py
@@ -8,7 +8,8 @@ from typing import TYPE_CHECKING, Any, cast
 import libs.global_value as g
 from integrations.base.interface import MessageParserDataMixin, MessageParserInterface
 from integrations.protocols import MsgData, PostData, StatusData
-from libs.domain.datamodels import ChannelType, CommandType, MessageStatus
+from libs.domain.datamodels import ChannelType, CommandType
+from libs.types import MessageStatus
 
 if TYPE_CHECKING:
     from integrations.slack.adapter import ServiceAdapter
