+++ integrations/standard_io/parser.py
@@ -7,7 +7,8 @@ from typing import Any, cast
 
 from integrations.base.interface import MessageParserDataMixin, MessageParserInterface
 from integrations.protocols import MsgData, PostData, StatusData
-from libs.domain.datamodels import ChannelType, MessageStatus
+from libs.domain.datamodels import ChannelType
+from libs.types import MessageStatus
 
 
 class MessageParser(MessageParserDataMixin, MessageParserInterface):
