+++ libs/data/modify.py
@@ -12,9 +12,9 @@ from typing import TYPE_CHECKING, Any, cast
 
 import libs.global_value as g
 from libs.data import lookup
-from libs.domain.datamodels import ActionStatus, MessageStatus
+from libs.domain.datamodels import ActionStatus
 from libs.functions import message
-from libs.types import StyleOptions
+from libs.types import MessageStatus, StyleOptions
 from libs.utils import dbutil, formatter
 from libs.utils.timekit import ExtendedDatetime as ExtDt
 from libs.utils.timekit import Format
