+++ libs/domain/datamodels.py
@@ -44,21 +44,6 @@ class CommandType(StrEnum):
     """未定義"""
 
 
-class MessageStatus(StrEnum):
-    """メッセージステータス"""
-
-    APPEND = "message_append"
-    """新規ポストイベント"""
-    CHANGED = "message_changed"
-    """編集イベント"""
-    DELETED = "message_deleted"
-    """削除イベント"""
-    DO_NOTHING = "do_nothing"
-    """何もしなくてよいイベント"""
-    UNDETERMINED = "undetermined"
-    """未定義状態"""
-
-
 class ActionStatus(StrEnum):
     """DBに対する操作"""
